Fix part of #4044, part of #1617: Replace proto formatting mechanism (#4045)
## Explanation
Fix part of #4044
Fix part of #1617
This PR makes #1543 obsolete once merged.
Originally copied from #2173 when it was in proof-of-concept form

This PR replaces the proto formatting mechanism (or, rather, makes it obsolete) by leveraging a feature of Bazel's proto library that allows the import prefix to be stripped. This ensures that the generated protos are treated as though they're in the same directory (which is how they were set up before per Gradle's requirement). The main benefit of this slightly cleaner approach is it allows the textproto conversion pipeline to work for protos that import other protos (otherwise the path doesn't match up, and it ends up with a Bazel error).

This required introducing a new proto library to ensure the prefix is properly stripped, and thus this PR introduces a regex check + test to force this check. As part of introducing this check, file matching & exemption checking now allows for more flexible regex (via checks).

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If this PR fixes part of an issue, prefix the title with "Fix part of #bugnum: ...".)
- [x] Any changes to [scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets) files have their rationale included in the PR explanation.
- [x] The PR follows the [style guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android Studio ([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers ([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- this only changes build infrastructure & has no direct impact on user interfaces.

Commit history:

* Copy proto-based changes from #2173.

* Post-merge fix.

* Add regex check, docs, and resolve TODOs.

This also changes regex handling in the check to be more generic for
better flexibility when matching files.

* Lint fix.

* Fix broken build.

* Post-merge fix.