SW translation fixes + better UI.
This fixes a bunch of Swahili string bugs by stripping out unnecessary
prefixed whitespace.

This also introduces a fast language switch bar between English and
Swahili. It's not currently tested or gated behind a platform flag, but
both of these will be changed if the team decides to proceed with this
solution.